made collabData a component
