+++ whiteboard-client/src/components/Collaborators/collabData.js
@@ -1,0 +1,8 @@
+  const collaborators = [
+    {id:1,name:"james",order:"first"},
+    {id:2,name:"Kamzzy",order:"second"},
+    {id:3,name:"Tamara",order:"third"},
+    {id:4,name:"Snow",order:"fourth"}
+  ]
+
+  export default collaborators;
